@@ -26,7 +26,7 @@
 		"theme": "dark"
 	},
 	"engines": {
-		"vscode": "^0.10.1"
+		"vscode": "^0.11.0"
 	},
 	"activationEvents": [
 		"onLanguage:dart",
@@ -143,7 +143,7 @@
 		"compile": "node ./node_modules/vscode/bin/compile -watch -p ./"
 	},
 	"devDependencies": {
-		"typescript": "^1.6.2",
-		"vscode": "0.10.x"
+		"typescript": "^1.7.3",
+		"vscode": "^0.11.10"
 	}
 }
