@@ -1,6 +1,8 @@
 "use strict";
 
-import vscode = require("vscode");
+// The module 'vscode' contains the VS Code extensibility API
+// Import the module and reference it with the alias vscode in your code below
+import * as vscode from 'vscode';
 import cp = require("child_process");
 
 import { LinterResult } from "./dartLintResult";
