@@ -1,5 +1,7 @@
 "use strict";
 
-import vscode = require("vscode");
+// The module 'vscode' contains the VS Code extensibility API
+// Import the module and reference it with the alias vscode in your code below
+import * as vscode from 'vscode';
 
 export const DART_MODE: vscode.DocumentFilter = { language: "dart", scheme: "file" }
